+++ server/src/main/java/com/server/server/data/RoadTrafficData.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class RoadTrafficData {
-    private int roadId;
+    private long roadId;
     private int userCount;
     private int maxLoad;
     private String status;
