+++ server/src/main/java/com/server/server/data/TrafficData.java
@@ -4,7 +4,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class TrafficData {
-    private int roadId;     // 外键
+    private long roadId;     // 外键
     private int userId;     // 外键
     private float speed;
     private String timestamp; // 可以用 java.util.Date 或 Timestamp 代替
